* feature: check saldo * feature: check mutasi * feature: buyback * enhance: data type in harga model

+++ microservice/check-mutasi-service/internal/models/transaksi.go
@@ -11,7 +11,7 @@ type TransaksiData struct {
 	Type         string  `json:"type" gorm:"type:varchar(20)"`
 	Gram         float64 `json:"gram" gorm:"type:decimal(12,2)"`
 	Saldo        float64 `json:"saldo" gorm:"type:decimal(12,2)"`
-	HargaTopup   float64 `json:"harga_topup" gorm:"type:integer;"`
-	HargaBuyback float64 `json:"harga_buyback" gorm:"type:integer"`
+	HargaTopup   int64   `json:"harga_topup" gorm:"type:integer;"`
+	HargaBuyback int64   `json:"harga_buyback" gorm:"type:integer"`
 	NoRek        string  `json:"no_rek" gorm:"type:varchar(15)"`
 }
